requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
aom=3.6.0=h6a678d5_0
astra-toolbox=2.1.3=py38h9d24096_0
asttokens=2.4.1=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
blas=1.0=mkl
blosc=1.21.3=h6a678d5_0
brotli=1.0.9=h5eee18b_8
brotli-bin=1.0.9=h5eee18b_8
brotli-python=1.0.9=py38h6a678d5_8
brunsli=0.1=h2531618_0
bzip2=1.0.8=h5eee18b_6
c-ares=1.19.1=h5eee18b_0
ca-certificates=2024.6.2=hbcca054_0
certifi=2024.6.2=py38h06a4308_0
cfitsio=3.470=h5893167_7
charls=2.2.0=h2531618_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=py38h06a4308_0
cloudpickle=2.2.1=py38h06a4308_0
comm=0.2.2=pyhd8ed1ab_0
contourpy=1.0.5=py38hdb19cb5_0
cuda-cudart=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.5.39=0
cuda-runtime=12.1.0=0
cuda-version=12.5=3
cycler=0.11.0=pyhd3eb1b0_0
cyrus-sasl=2.1.28=h52b45da_1
cytoolz=0.12.2=py38h5eee18b_0
dask-core=2023.4.1=py38h06a4308_0
dav1d=1.2.1=h5eee18b_0
dbus=1.13.18=hb2f20db_0
debugpy=1.6.7=py38h6a678d5_0
decorator=5.1.1=pyhd8ed1ab_0
executing=2.0.1=pyhd8ed1ab_0
expat=2.6.2=h6a678d5_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py38h06a4308_0
fontconfig=2.14.1=h4c34cd2_2
fonttools=4.51.0=py38h5eee18b_0
freetype=2.12.1=h4a9f257_0
fsspec=2024.3.1=py38h06a4308_0
giflib=5.2.1=h5eee18b_3
glib=2.78.4=h6a678d5_0
glib-tools=2.78.4=h6a678d5_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py38heeb90bb_0
gnutls=3.6.15=he1e5248_0
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
h5py=3.11.0=py38hbe37b52_0
hdf5=1.12.1=h2b7332f_3
icu=73.1=h6a678d5_0
idna=3.7=py38h06a4308_0
imagecodecs=2023.1.23=py38hc4b7b5f_0
imageio=2.33.1=py38h06a4308_0
importlib-metadata=8.0.0=pyha770c72_0
importlib_metadata=8.0.0=hd8ed1ab_0
importlib_resources=6.1.1=py38h06a4308_1
intel-openmp=2023.1.0=hdb19cb5_46306
ipykernel=6.29.5=pyh3099207_0
ipython=8.12.2=pyh41d4057_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.4=py38h06a4308_0
jpeg=9e=h5eee18b_1
jupyter_client=8.6.2=pyhd8ed1ab_0
jupyter_core=5.7.2=py38h578d9bd_0
jxrlib=1.1=h7b6447c_2
kiwisolver=1.4.4=py38h6a678d5_0
krb5=1.20.1=h143b758_1
lame=3.100=h7b6447c_0
lazy_loader=0.4=py38h06a4308_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libaec=1.0.4=he6710b0_1
libastra=2.1.3=cuda_12.1_0
libavif=0.11.1=h5eee18b_0
libblas=3.9.0=1_h86c2bf4_netlib
libbrotlicommon=1.0.9=h5eee18b_8
libbrotlidec=1.0.9=h5eee18b_8
libbrotlienc=1.0.9=h5eee18b_8
libcblas=3.9.0=5_h92ddd45_netlib
libclang=14.0.6=default_hc6dbbc7_1
libclang13=14.0.6=default_he11475f_1
libcublas=12.1.0.26=0
libcufft=11.0.2.4=0
libcufile=1.10.0.4=0
libcups=2.4.2=h2d74bed_1
libcurand=10.3.6.39=0
libcurl=8.7.1=h251f7ec_0
libcusolver=11.4.4.55=0
libcusparse=12.0.2.55=0
libdeflate=1.17=h5eee18b_1
libedit=3.1.20230828=h5eee18b_0
libev=4.33=h7f8727e_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=14.1.0=h77fa898_0
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libglib=2.78.4=hdc74915_0
libgomp=14.1.0=h77fa898_0
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
liblapack=3.9.0=5_h92ddd45_netlib
libllvm14=14.0.6=hdb19cb5_3
libnghttp2=1.57.0=h2d74bed_0
libnpp=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjpeg=12.1.1.14=0
libpng=1.6.39=h5eee18b_0
libpq=12.17=hdbd6064_0
libsodium=1.0.18=h36c2ea0_1
libssh2=1.11.0=h251f7ec_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp-base=1.3.2=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=h5eee18b_1
libxml2=2.10.4=hfdd30dd_2
libzopfli=1.0.3=he6710b0_0
llvm-openmp=14.0.6=h9e868ea_0
locket=1.0.0=py38h06a4308_0
lz4-c=1.9.4=h6a678d5_1
markupsafe=2.1.3=py38h5eee18b_0
matplotlib=3.7.2=py38h06a4308_0
matplotlib-base=3.7.2=py38h1128e8f_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py38h5eee18b_1
mkl_fft=1.3.8=py38h5eee18b_0
mkl_random=1.2.4=py38hdb19cb5_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py38h06a4308_0
mysql=5.7.24=h721c034_2
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py38h06a4308_0
numpy=1.24.3=py38hf6e8229_1
numpy-base=1.24.3=py38h060ed82_1
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h9ca470c_1
openssl=3.3.1=h4ab18f5_1
packaging=23.2=py38h06a4308_0
parso=0.8.4=pyhd8ed1ab_0
partd=1.4.1=py38h06a4308_0
pcre2=10.42=hebb0a14_1
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.3.0=py38h5eee18b_0
pip=24.0=py38h06a4308_0
platformdirs=4.2.2=pyhd8ed1ab_0
ply=3.11=py38_0
pooch=1.7.0=py38h06a4308_0
prompt-toolkit=3.0.47=pyha770c72_0
prompt_toolkit=3.0.47=hd8ed1ab_0
psutil=6.0.0=py38hfb59056_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pygments=2.18.0=pyhd8ed1ab_0
pyparsing=3.0.9=py38h06a4308_0
pyqt=5.15.10=py38h6a678d5_0
pyqt5-sip=12.13.0=py38h5eee18b_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.19=h955ad1f_0
python-dateutil=2.9.0post0=py38h06a4308_2
python_abi=3.8=2_cp38
pytorch=2.3.1=py3.8_cuda12.1_cudnn8.9.2_0
pytorch-cuda=12.1=ha16c6d3_5
pytorch-mutex=1.0=cuda
pywavelets=1.4.1=py38h5eee18b_0
pyyaml=6.0.1=py38h5eee18b_0
pyzmq=25.1.2=py38h6a678d5_0
qt-main=5.15.2=h53bd1ea_10
readline=8.2=h5eee18b_0
requests=2.32.2=py38h06a4308_0
scikit-image=0.20.0=py38h6a678d5_0
scipy=1.8.1=py38h1ee437e_0
setuptools=69.5.1=py38h06a4308_0
sip=6.7.12=py38h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.10=h6a678d5_1
sqlite=3.45.3=h5eee18b_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.12=py38h06a4308_0
tbb=2021.8.0=hdb19cb5_0
tifffile=2023.4.12=py38h06a4308_0
tk=8.6.14=h39e8969_0
tomli=2.0.1=py38h06a4308_0
toolz=0.12.0=py38h06a4308_0
torchaudio=2.3.1=py38_cu121
torchtriton=2.3.1=py38
torchvision=0.18.1=py38_cu121
tornado=6.4.1=py38h5eee18b_0
traitlets=5.14.3=pyhd8ed1ab_0
typing_extensions=4.11.0=py38h06a4308_0
unicodedata2=15.1.0=py38h5eee18b_0
urllib3=2.2.2=py38h06a4308_0
wcwidth=0.2.13=pyhd8ed1ab_0
wheel=0.43.0=py38h06a4308_0
xz=5.4.6=h5eee18b_1
yaml=0.2.5=h7b6447c_0
zeromq=4.3.5=h6a678d5_0
zfp=1.0.0=h6a678d5_0
zipp=3.17.0=py38h06a4308_0
zlib=1.2.13=h5eee18b_1
zstd=1.5.5=hc292b87_2
